Add unit tests using `test_cases` & bump dependencies versions
