+++ lib/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cairo-lang-sierra = "~2.6.4"
+cairo-lang-sierra = "~2.7.1"
 colored = "2.1.0"
 graphviz-rust = "0.9.0"
 hex = "0.4.3"
@@ -16,9 +16,10 @@ regex = "1.10.4"
 reqwest = { version = "0.12.4", features = ["json"] }
 serde = "1.0.198"
 serde_json = "1.0.116"
+test-case = "3.3.1"
 tokio = { version="^1.37.0", features = ["full"] }
 z3 = "0.12.1"
 
 [dev-dependencies]
 serde_json = "1.0.116"
-cairo-lang-starknet-classes = "~2.6.4"
+cairo-lang-starknet-classes = "~2.7.1"
