+++ lib/tests/detectors.rs
