+++ sierra-decompiler/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cairo-lang-sierra = "~2.6.4" 
-cairo-lang-starknet-classes = "~2.6.4" 
+cairo-lang-sierra = "~2.7.1" 
+cairo-lang-starknet-classes = "~2.7.1" 
 clap = { version = "4.0.0-rc.1", features = [ "derive" ] }
 serde_json = "1.0.116"
 sierra-analyzer-lib = { path = "../lib" }
